requirements.txt
flask>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.13.0
openai>=1.40.0
python-dotenv>=1.0.0
pytest>=8.0.0
reportlab>=4.0.0
pillow>=10.0.0
